API will now append CreatedById to parcel evals and fiscals. Lifted Property dialogs into own file. Updating general information for parcels and buildings works but needs more testing.
