+++ react-app/src/components/property/PropertyForms.tsx
@@ -1,24 +1,13 @@
-import { Parcel } from '@/hooks/api/useParcelsApi';
 import { Box, Grid, InputAdornment, Tooltip, Typography } from '@mui/material';
-import React, { useEffect } from 'react';
-import { FormProvider, useForm } from 'react-hook-form';
 import TextFormField from '../form/TextFormField';
 import AutocompleteFormField from '../form/AutocompleteFormField';
 import SelectFormField, { ISelectMenuItem } from '../form/SelectFormField';
 import { Help } from '@mui/icons-material';
 import { LookupObject } from '@/hooks/api/useLookupApi';
 import DateFormField from '../form/DateFormField';
-import useDataLoader from '@/hooks/useDataLoader';
-import usePimsApi from '@/hooks/usePimsApi';
-import ConfirmDialog from '../dialog/ConfirmDialog';
-import { Building } from '@/hooks/api/useBuildingsApi';
-import dayjs from 'dayjs';
+import React from 'react';
 
 export type PropertyType = 'Building' | 'Parcel';
-
-interface IParcelInformationEditDialog {
-  initialValues: Parcel;
-}
 
 interface IParcelInformationForm {
   classificationOptions: ISelectMenuItem[];
@@ -232,181 +221,5 @@ export const BuildingInformationForm = (props: IBuildingInformationForm) => {
         </Grid>
       </Grid>
     </>
-  );
-};
-
-interface IParcelInformationEditDialog {
-  initialValues: Parcel;
-  open: boolean;
-  onCancel: () => void;
-}
-
-export const ParcelInformationEditDialog = (props: IParcelInformationEditDialog) => {
-  const { initialValues } = props;
-
-  const api = usePimsApi();
-
-  const { data: adminAreasData, loadOnce: adminLoad } = useDataLoader(
-    api.administrativeAreas.getAdministrativeAreas,
-  );
-  const { data: classificationData, loadOnce: classificationLoad } = useDataLoader(
-    api.lookup.getClassifications,
-  );
-
-  adminLoad();
-  classificationLoad();
-
-  const infoFormMethods = useForm({
-    defaultValues: {
-      NotOwned: true,
-      Address1: '',
-      PIN: null,
-      PID: null,
-      Postal: '',
-      AdministrativeAreaId: null,
-      LandArea: null,
-      IsSensitive: false,
-      ClassificationId: null,
-      Description: '',
-    },
-  });
-
-  useEffect(() => {
-    infoFormMethods.reset({
-      NotOwned: initialValues?.NotOwned,
-      Address1: initialValues?.Address1,
-      PIN: initialValues?.PIN,
-      PID: initialValues?.PID,
-      Postal: initialValues?.Postal,
-      AdministrativeAreaId: initialValues?.AdministrativeAreaId,
-      LandArea: initialValues?.LandArea,
-      IsSensitive: initialValues?.IsSensitive,
-      ClassificationId: initialValues?.ClassificationId,
-      Description: initialValues?.Description,
-    });
-  }, [initialValues]);
-  return (
-    <ConfirmDialog
-      title={'Edit parcel information'}
-      open={props.open}
-      onConfirm={async () => console.log(infoFormMethods.getValues())}
-      onCancel={async () => props.onCancel()}
-    >
-      <FormProvider {...infoFormMethods}>
-        <Box display={'flex'} flexDirection={'column'} gap={'1rem'}>
-          <GeneralInformationForm
-            propertyType={'Parcel'}
-            adminAreas={
-              adminAreasData?.map((admin) => ({ label: admin.Name, value: admin.Id })) ?? []
-            }
-          />
-          <ParcelInformationForm
-            classificationOptions={
-              classificationData?.map((classif) => ({
-                label: classif.Name,
-                value: classif.Id,
-              })) ?? []
-            }
-          />
-        </Box>
-      </FormProvider>
-    </ConfirmDialog>
-  );
-};
-
-interface IBuildingInformationEditDialog {
-  initialValues: Building;
-  open: boolean;
-  onCancel: () => void;
-}
-
-export const BuildingInformationEditDialog = (props: IBuildingInformationEditDialog) => {
-  const api = usePimsApi();
-
-  const { data: adminAreasData, loadOnce: adminLoad } = useDataLoader(
-    api.administrativeAreas.getAdministrativeAreas,
-  );
-  const { data: classificationData, loadOnce: classificationLoad } = useDataLoader(
-    api.lookup.getClassifications,
-  );
-  const { data: predominateUseData, loadOnce: predominateUseLoad } = useDataLoader(
-    api.lookup.getPredominateUses,
-  );
-  const { data: constructionTypeData, loadOnce: constructionLoad } = useDataLoader(
-    api.lookup.getConstructionTypes,
-  );
-
-  adminLoad();
-  classificationLoad();
-  predominateUseLoad();
-  constructionLoad();
-
-  const { initialValues, open, onCancel } = props;
-  const infoFormMethods = useForm({
-    defaultValues: {
-      NotOwned: true,
-      Address1: '',
-      PIN: null,
-      PID: null,
-      Postal: '',
-      AdministrativeAreaId: null,
-      LandArea: null,
-      IsSensitive: false,
-      ClassificationId: null,
-      Description: '',
-      Name: '',
-      BuildingPredominateUseId: null,
-      BuildingConstructionTypeId: null,
-      TotalArea: null,
-      RentableArea: null,
-      BuildingTenancy: '',
-      BuildingTenancyUpdatedOn: dayjs(),
-    },
-  });
-
-  useEffect(() => {
-    infoFormMethods.reset({
-      Address1: initialValues?.Address1,
-      PIN: initialValues?.PIN,
-      PID: initialValues?.PID,
-      Postal: initialValues?.Postal,
-      AdministrativeAreaId: initialValues?.AdministrativeAreaId,
-      LandArea: initialValues?.TotalArea,
-      IsSensitive: initialValues?.IsSensitive,
-      ClassificationId: initialValues?.ClassificationId,
-      Description: initialValues?.Description,
-      Name: initialValues?.Name,
-      BuildingPredominateUseId: initialValues?.BuildingPredominateUseId,
-      BuildingConstructionTypeId: initialValues?.BuildingConstructionTypeId,
-      TotalArea: initialValues?.TotalArea,
-      RentableArea: initialValues?.RentableArea,
-      BuildingTenancy: initialValues?.BuildingTenancy,
-      BuildingTenancyUpdatedOn: initialValues?.BuildingTenancyUpdatedOn,
-    });
-  }, [initialValues]);
-
-  return (
-    <ConfirmDialog
-      title={'Edit building information'}
-      open={open}
-      onConfirm={async () => console.log(infoFormMethods.getValues())}
-      onCancel={async () => onCancel()}
-    >
-      <FormProvider {...infoFormMethods}>
-        <Box display={'flex'} flexDirection={'column'} gap={'1rem'}>
-          <GeneralInformationForm
-            propertyType={'Building'}
-            adminAreas={
-              adminAreasData?.map((admin) => ({ label: admin.Name, value: admin.Id })) ?? []
-            }
-          />
-          <BuildingInformationForm
-            classificationOptions={classificationData}
-            constructionOptions={constructionTypeData}
-            predominateUseOptions={predominateUseData}
-          />
-        </Box>
-      </FormProvider>
-    </ConfirmDialog>
   );
 };
